server/requirements.txt
flask~=2.0.1
sqlalchemy~=1.4.21
flask-sqlalchemy
protobuf
absl-py
ecdsa
